write report error fix in cucumber parallelism
