+++ lib/runner/cli/cli.js
@@ -280,7 +280,8 @@ class CliRunner {
       NightwatchEventHub.runner = this.testRunner.type;
       NightwatchEventHub.isAvailable = true;
       
-      const {globals} = this.test_settings;
+      const {globals, output_folder} = this.test_settings;
+      NightwatchEventHub.output_folder = output_folder;
       const {plugins} = this.globals;
   
       if (Utils.isFunction(globals.registerEventHandlers)) {
