+++ lib/runner/eventHub.js
@@ -1,5 +1,7 @@
+const fs = require('fs');
+const path = require('path');
 const EventEmitter = require('events');
-const {Logger} = require('../utils');
+const {Logger, createFolder} = require('../utils');
 
 class NightwatchEventHub extends EventEmitter {
   emit(eventName, data) {
@@ -9,6 +11,17 @@ class NightwatchEventHub extends EventEmitter {
       } catch (err) {
         Logger.error(err);
       }
+    }
+
+    if (eventName === 'TestFinished' && this.output_folder && this.runner === 'cucumber') {
+      let {output_folder} = this;
+      output_folder = path.join(output_folder, 'cucumber');
+      const filename = path.join(output_folder, 'cucumber-report.json');
+
+      this.writeReportFile(filename, JSON.stringify(data.report, null, 2), true, output_folder)
+        .then(_ => {
+          Logger.info(Logger.colors.stack_trace(`Wrote JSON report file to: ${path.resolve(filename)}`));
+        });
     }
   }
 
@@ -26,6 +39,21 @@ class NightwatchEventHub extends EventEmitter {
 
   get runner() {
     return this.runnerType;
+  }
+
+  writeReportFile(filename, rendered, shouldCreateFolder, output_folder) {
+    return (shouldCreateFolder ? createFolder(output_folder) : Promise.resolve())
+      .then(() => {
+        return new Promise((resolve, reject) => {
+          fs.writeFile(filename, rendered, function(err) {
+            if (err) {
+              return reject(err);
+            }
+
+            resolve();
+          });
+        });
+      });
   }
 }
