+++ lib/runner/test-runners/cucumber/_setup_cucumber_runner.js
@@ -52,8 +52,6 @@ Before(function({pickle}) {
     });
   }
 
-  NightwatchFormat.setClient(this.client);
-
   // eslint-disable-next-line
   console.log('\n');
