+++ lib/runner/test-runners/cucumber/nightwatch-format.js
@@ -1,7 +1,5 @@
-const fs = require('fs');
-const path = require('path');
 const Utils = require('../../../utils');
-const {Logger, createFolder} = Utils;
+const {Logger} = Utils;
 const {Formatter} = require('@cucumber/cucumber');
 const lodashCloneDeep = require('lodash.clonedeep');
 
@@ -30,10 +28,6 @@ module.exports = class NightwatchFormatter extends Formatter {
 
       this.reportHandler(envelopeCopy);
     });
-  }
-
-  static setClient(client){
-    NightwatchFormatter.client = client;
   }
 
   static setCapabilities({sessionId, capabilities}) {
@@ -103,30 +97,6 @@ module.exports = class NightwatchFormatter extends Formatter {
       envelope: result,
       report: this.report
     });
-
-    let {output_folder} = NightwatchFormatter.client.settings;
-    output_folder = path.join(output_folder, 'cucumber');
-    const filename = path.join(output_folder, 'cucumber-report.json');
-
-    this.writeReportFile(filename, JSON.stringify(this.report, null, 2), true, output_folder)
-      .then(_ => {
-        Logger.info(Logger.colors.stack_trace(`Wrote JSON report file to: ${path.resolve(filename)}`));
-      });
-  }
-
-  writeReportFile(filename, rendered, shouldCreateFolder, output_folder) {
-    return (shouldCreateFolder ? createFolder(output_folder) : Promise.resolve())
-      .then(() => {
-        return new Promise((resolve, reject) => {
-          fs.writeFile(filename, rendered, function(err) {
-            if (err) {
-              return reject(err);
-            }
-
-            resolve();
-          });
-        });
-      });
   }
 
   onTestRunStarted(result) {
@@ -135,8 +105,7 @@ module.exports = class NightwatchFormatter extends Formatter {
     NightwatchEventHub.emit(TestStarted, {
       envelope: result,
       report: this.report
-    }
-    );
+    });
   }
 
   onTestStepFinished(result) {
